+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Switched to YAML for the `.ig_markets` config files and added support for storing multiple authentication profiles in
   the config files that can then be selected on the command-line using the new `--profile NAME` argument
+- Added `IGMarkets::DealingPlatform#client_account_summary`
 - Report the HTTP code when a request fails
 
 ### 0.15 — July 11, 2016